e2e: Rename e2e package to testing
Preparing for adding validation package for decoupling command
management and accessing actual clusters.
